requirements.txt
alethic-ism-core
alethic-ism-db
python-dotenv
psycopg2
python-multipart
restrictedpython
#redis
pyarrow
pyyaml
fastapi
pydantic
nats-py
mako
pyjwt
datasets
uvicorn
firebase-admin
starlette
openpyxl